+++ Scripts/PowerShell/Rename-IntuneDevice/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Rename-IntuneComputer release history
 
+## 1.1.0 - 2024-01-15
+
+* Fixed missing `DeviceName` parameter.
+* Fixed default values for script parameters.
+* Removed obsolete `UserPrincipalName` parameter.
+* Replaced `${ScriptSection}` with `$script:ScriptSection` for `Write-Log` in main script.
+
 ## 1.0.0 - 2024-01-12
 
 ### First version